+++ lib/execution_engine2/sdk/EE2Runjob.py
@@ -7,8 +7,7 @@ the logic to retrieve info needed by the runnner to start the job
 import os
 import time
 from collections import Counter
-from enum import Enum
-from typing import Optional, Dict, NamedTuple, Union, List, Any
+from typing import Optional, Dict, Union, List, Any
 
 from execution_engine2.db.models.models import (
     Job,
@@ -45,39 +44,11 @@ from execution_engine2.utils.job_requirements_resolver import (
     DEBUG_MODE,
 )
 from execution_engine2.utils.job_requirements_resolver import RequirementsType
+from execution_engine2.sdk.EE2JobSubmissionExecutor import EE2JobSubmissionExecutor
 
-_JOB_REQUIREMENTS = "job_reqs"
-_JOB_REQUIREMENTS_INCOMING = "job_requirements"
-_SCHEDULER_REQUIREMENTS = "scheduler_requirements"
-_META = "meta"  # narrative_cell_info
-_APP_PARAMS = "params"  # application parameters
-_REQUIREMENTS_LIST = "requirements_list"
-_METHOD = "method"
-_APP_ID = "app_id"
-_BATCH_ID = "batch_id"
-_PARENT_JOB_ID = "parent_job_id"
-_PARENT_RETRY_JOB_ID = "retry_parent"
-_RETRY_IDS = "retry_ids"
-_WORKSPACE_ID = "wsid"
-_SOURCE_WS_OBJECTS = "source_ws_objects"
-_SERVICE_VER = "service_ver"
-
-
-class JobPermissions(Enum):
-    READ = "r"
-    WRITE = "w"
-    NONE = "n"
-
-
-class PreparedJobParams(NamedTuple):
-    params: dict
-    job_id: str
-
-
-class JobIdPair(NamedTuple):
-    job_id: str
-    scheduler_id: str
-
+from sdk.EE2RunjobConstants import _JOB_REQUIREMENTS, _JOB_REQUIREMENTS_INCOMING, _SCHEDULER_REQUIREMENTS, _META, \
+    _APP_PARAMS, _REQUIREMENTS_LIST, _METHOD, _APP_ID, _BATCH_ID, _PARENT_JOB_ID, _PARENT_RETRY_JOB_ID, _WORKSPACE_ID, \
+    _SOURCE_WS_OBJECTS, _SERVICE_VER, JobPermissions
 
 from typing import TYPE_CHECKING
 
@@ -86,10 +57,12 @@ if TYPE_CHECKING:
 
 
 class EE2RunJob:
-    def __init__(self, sdkmr):
+    def __init__(self, sdkmr: SDKMethodRunner, jse: EE2JobSubmissionExecutor):
         self.sdkmr = sdkmr  # type: SDKMethodRunner
+        # Looks like this feature was deleted
         self.override_clientgroup = os.environ.get("OVERRIDE_CLIENT_GROUP", None)
         self.logger = self.sdkmr.get_logger()
+        self.jse = jse
 
     def _init_job_rec(
         self, user_id: str, params: Dict, save: bool = True
@@ -216,42 +189,7 @@ class EE2RunJob:
                     f"User {self.sdkmr.user_id} doesn't have permission to run jobs in workspace {bad_ws}."
                 )
 
-    def _finish_created_job(
-        self, job_id, exception, error_code=None, error_message=None
-    ):
-        """
-        :param job_id:
-        :param exception:
-        :param error_code:
-        :param error_message:
-        :return:
-        """
-        if error_message is None:
-            error_message = f"Cannot submit to condor {exception}"
-        if error_code is None:
-            error_code = ErrorCode.job_crashed.value
 
-        # TODO Better Error Parameter? A Dictionary?
-        self.sdkmr.finish_job(
-            job_id=job_id,
-            error_message=error_message,
-            error_code=error_code,
-            error=f"{exception}",
-        )
-
-    def _generate_job_submission_params(self, job_id, params):
-        return JobSubmissionParameters(
-            job_id,
-            AppInfo(params[_METHOD], params.get(_APP_ID)),
-            params[_JOB_REQUIREMENTS],
-            UserCreds(self.sdkmr.get_user_id(), self.sdkmr.get_token()),
-            # a job should have a parent ID or a batch ID or nothing, but never both
-            # Do we want to distinguish between the two cases in the sub params?
-            # It's informational only for Condor
-            parent_job_id=params.get(_BATCH_ID) or params.get(_PARENT_JOB_ID),
-            wsid=params.get(_WORKSPACE_ID),
-            source_ws_objects=params.get(_SOURCE_WS_OBJECTS),
-        )
 
     def _prepare_to_run(self, params, concierge_params=None) -> JobSubmissionParameters:
         """
@@ -262,7 +200,9 @@ class EE2RunJob:
         self.logger.debug(
             f"User {self.sdkmr.get_user_id()} attempting to run job {params[_METHOD]} {params}"
         )
-        return self._generate_job_submission_params(job_id, params)
+        return self.jse.generate_job_submission_params(job_id, params)
+
+
 
     def _run_multiple(self, runjob_params):
         """
@@ -278,104 +218,16 @@ class EE2RunJob:
             )
         job_ids = self.sdkmr.save_jobs(job_records)
 
-        # Generate job submission params
-        job_submission_params = []
-        for i, job_id in enumerate(job_ids):
-            job_submission_params.append(
-                self._generate_job_submission_params(job_id, runjob_params[i])
-            )
-            assert job_id == job_submission_params[i].job_id
-
-        # Takes 2.5200018882751465 for 100 records, can shave off 2.5 secs by making this async
-        for job_id in job_ids:
-            self.sdkmr.get_kafka_client().send_kafka_message(
-                message=KafkaCreateJob(
-                    job_id=str(job_id), user=self.sdkmr.get_user_id()
-                )
-            )
-        # Submit to Condor
-        try:
-            submission_ids = self._submit_multiple(job_submission_params)
-            return submission_ids
-        except Exception as e:
-            self._abort_multiple_jobs(job_ids)
-            raise e
-
-    def _update_to_queued_multiple(self, job_ids, scheduler_ids):
-        """
-        This is called during job submission. If a job is terminated during job submission,
-        we have the chance to re-issue a termination and remove the job from the Job Queue
-        """
-        if len(job_ids) != len(scheduler_ids):
-            raise Exception(
-                "Need to provide the same amount of job ids and scheduler_ids"
-            )
-        jobs_to_update = list(map(JobIdPair, job_ids, scheduler_ids))
-        self.sdkmr.get_mongo_util().update_jobs_to_queued(jobs_to_update)
-        jobs = self.sdkmr.get_mongo_util().get_jobs(job_ids)
-
-        for job in jobs:
-            job_id = str(job.id)
-            if job.status == Status.queued.value:
-                self.sdkmr.get_kafka_client().send_kafka_message(
-                    message=KafkaQueueChange(
-                        job_id=job_id,
-                        new_status=Status.queued.value,
-                        previous_status=Status.created.value,  # TODO maybe change this to allow for estimating jobs
-                        scheduler_id=job.scheduler_id,
-                    )
-                )
-            elif job.status == Status.terminated.value:
-                # Remove from the queue, now that the scheduler_id is available
-                # The job record doesn't actually get updated in the db a 2nd time, and this TerminatedCode is only
-                # used by the initial transition to Terminated
-                self._safe_cancel(job_id, TerminatedCode.terminated_by_user)
-
-    def _submit_multiple(self, job_submission_params):
-        """
-        Submit multiple jobs. If any of the submissions are a failure, raise exception in order
-        to fail all submitted jobs, rather than allowing the submissions to continue
-        """
-        begin = time.time()
-        job_ids = []
-        condor_job_ids = []
-        for job_submit_param in job_submission_params:
-            job_id = job_submit_param.job_id
-            job_ids.append(job_id)
-            try:
-                submission_info = self.sdkmr.get_condor().run_job(
-                    params=job_submit_param
-                )
-                condor_job_id = submission_info.clusterid
-            except Exception as e:
-                self.logger.error(e)
-                self._finish_created_job(job_id=job_id, exception=e)
-                raise e
-
-            if submission_info.error is not None and isinstance(
-                submission_info.error, Exception
-            ):
-                self._finish_created_job(exception=submission_info.error, job_id=job_id)
-                raise submission_info.error
-            if condor_job_id is None:
-                error_msg = (
-                    "Condor job not run, and error not found. Something went wrong"
-                )
-                self._finish_created_job(
-                    job_id=job_id, exception=RuntimeError(error_msg)
-                )
-                raise RuntimeError(error_msg)
-            condor_job_ids.append(condor_job_id)
-
-        self.logger.error(f"It took {time.time() - begin} to submit jobs to condor")
-        # It took 4.836009502410889 to submit jobs to condor
-
-        update_time = time.time()
-        self._update_to_queued_multiple(job_ids=job_ids, scheduler_ids=condor_job_ids)
-        # It took 1.9239885807037354 to update jobs
-        self.logger.error(f"It took {time.time() - update_time} to update jobs ")
+        EE2JobSubmissionExecutor(self).async_submit_multiple(runjob_params=runjob_params, job_ids=job_ids)
 
         return job_ids
+
+
+
+
+
+
+
 
     def _run(self, params):
         job_params = self._prepare_to_run(params=params)
@@ -403,19 +255,6 @@ class EE2RunJob:
 
         return job_id
 
-    def _abort_multiple_jobs(self, job_ids):
-        """
-        Cancel a list of child jobs, and their child jobs
-        """
-        for job_id in job_ids:
-            try:
-                self.sdkmr.cancel_job(
-                    job_id=job_id,
-                    terminated_code=TerminatedCode.terminated_by_batch_abort.value,
-                )
-            except Exception as e:
-                # TODO Maybe add a retry here?
-                self.logger.error(f"Couldn't cancel child job {e}")
 
     def _create_batch_job(self, wsid, meta):
         """
@@ -462,7 +301,7 @@ class EE2RunJob:
         try:
             self.sdkmr.add_child_jobs(batch_job=batch_job, child_jobs=child_jobs)
         except Exception as e:
-            self._abort_multiple_jobs(child_jobs)
+            self.jse.abort_multiple_jobs(child_jobs)
             raise e
 
         return child_jobs
@@ -622,15 +461,7 @@ class EE2RunJob:
     def _retryable(status: str):
         return status in [Status.terminated.value, Status.error.value]
 
-    def _safe_cancel(
-        self,
-        job_id: str,
-        terminated_code: TerminatedCode,
-    ):
-        try:
-            self.sdkmr.cancel_job(job_id=job_id, terminated_code=terminated_code.value)
-        except Exception as e:
-            self.logger.error(f"Couldn't cancel {job_id} due to {e}")
+
 
     def _db_update_failure(
         self, job_that_failed_operation: str, job_to_abort: str, exception: Exception
@@ -641,9 +472,10 @@ class EE2RunJob:
             f"Couldn't update job record:{job_that_failed_operation} during retry. Aborting:{job_to_abort}"
             f" Exception:{exception} "
         )
-        self._safe_cancel(
+        self.sdkmr.cancel_job(
             job_id=job_to_abort,
             terminated_code=TerminatedCode.terminated_by_server_failure,
+            ignore_exceptions=True
         )
         # TODO Maybe move this log into multiple so not multiple error messages are generated
         self.logger.error(msg, exc_info=True, stack_info=True)
